Fix null pointer bugs

+++ src/main/java/py/org/fundacionparaguaya/pspserver/surveys/services/impl/SnapshotServiceImpl.java
@@ -28,6 +28,7 @@ import py.org.fundacionparaguaya.pspserver.surveys.dtos.SnapshotIndicators;
 import py.org.fundacionparaguaya.pspserver.surveys.dtos.SurveyData;
 import py.org.fundacionparaguaya.pspserver.surveys.entities.SnapshotEconomicEntity;
 import py.org.fundacionparaguaya.pspserver.surveys.entities.SnapshotIndicatorEntity;
+import py.org.fundacionparaguaya.pspserver.surveys.entities.SnapshotIndicatorPriorityEntity;
 import py.org.fundacionparaguaya.pspserver.surveys.entities.SurveyEntity;
 import py.org.fundacionparaguaya.pspserver.surveys.enums.SurveyStoplightEnum;
 import py.org.fundacionparaguaya.pspserver.surveys.mapper.SnapshotEconomicMapper;
@@ -147,46 +148,57 @@ public class SnapshotServiceImpl implements SnapshotService {
 
         SnapshotIndicators toRet = new SnapshotIndicators();
 
-        SnapshotEconomicEntity originalSnapshot = economicRepository.findOne(snapshotId);
+        SnapshotEconomicEntity originalSnapshot = economicRepository
+                .findOne(snapshotId);
 
-        SurveyEntity survey = surveyRepository.getOne(originalSnapshot.getSurveyDefinition().getId());
-        List<String> indicatorGroup = survey.getSurveyDefinition().getSurveyUISchema().getGroupIndicators();
+        if (originalSnapshot!=null) {
 
-        List<String> order = survey.getSurveyDefinition().getSurveyUISchema().getUiOrder().stream()
-                .filter(field -> indicatorGroup.contains(field)).collect(Collectors.toList());
+            SurveyEntity survey = surveyRepository.getOne(originalSnapshot
+                    .getSurveyDefinition().getId());
+            List<String> indicatorGroup = survey.getSurveyDefinition()
+                    .getSurveyUISchema().getGroupIndicators();
 
-        List<SnapshotIndicatorPriority> priorities = priorityService
-                .getSnapshotIndicatorPriorityList(originalSnapshot.getSnapshotIndicator().getId());
-        toRet.setIndicatorsPriorities(priorities);
+            List<String> order = survey.getSurveyDefinition()
+                    .getSurveyUISchema().getUiOrder().stream()
+                    .filter(field -> indicatorGroup.contains(field))
+                    .collect(Collectors.toList());
 
-        SurveyData indicators = indicatorMapper.entityToDto(originalSnapshot.getSnapshotIndicator());
-        List<SurveyData> indicatorsToRet = new ArrayList<>();
-        if (indicatorGroup != null && !indicatorGroup.isEmpty() && order != null && !order.isEmpty()) {
+            List<SnapshotIndicatorPriority> priorities = priorityService
+                    .getSnapshotIndicatorPriorityList(originalSnapshot.
+                            getSnapshotIndicator().getId());
+            toRet.setIndicatorsPriorities(priorities);
 
-            order.forEach(indicator -> {
-                if (indicators.containsKey(indicator)) {
-                    SurveyData sd = new SurveyData();
-                    sd.put(INDICATOR_NAME, getNameFromCamelCase(indicator));
-                    sd.put(INDICATOR_VALUE, indicators.get(indicator));
-                    
-                    countIndicators(toRet, sd.get(INDICATOR_VALUE));
-                    indicatorsToRet.add(sd);
-                }
-            });
+            SurveyData indicators = indicatorMapper
+                    .entityToDto(originalSnapshot.getSnapshotIndicator());
+            List<SurveyData> indicatorsToRet = new ArrayList<>();
+            if (indicatorGroup != null && !indicatorGroup.isEmpty()
+                    && order != null && !order.isEmpty()) {
 
+                order.forEach(indicator -> {
+                    if (indicators.containsKey(indicator)) {
+                        SurveyData sd = new SurveyData();
+                        sd.put(INDICATOR_NAME, getNameFromCamelCase(indicator));
+                        sd.put(INDICATOR_VALUE, indicators.get(indicator));
+                        countIndicators(toRet, sd.get(INDICATOR_VALUE));
+                        indicatorsToRet.add(sd);
+                    }
+                });
+
+            }
+
+            toRet.setIndicatorsSurveyData(indicatorsToRet);
+            toRet.setCreatedAt(originalSnapshot.getCreatedAtAsISOString());
+            toRet.setSnapshotIndicatorId(originalSnapshot
+                    .getSnapshotIndicator().getId());
+            toRet.setSnapshotEconomicId(originalSnapshot.getId());
+            toRet.setSurveyId(originalSnapshot.getSurveyDefinition().getId());
+
+            //set family for information purpose
+            Long familyId = originalSnapshot.getFamily().getFamilyId();
+            toRet.setFamilyId(familyId);
+            toRet.setFamily(familyService.getFamilyById(familyId));
         }
 
-        toRet.setIndicatorsSurveyData(indicatorsToRet);
-        toRet.setCreatedAt(originalSnapshot.getCreatedAtAsISOString());
-        toRet.setSnapshotIndicatorId(originalSnapshot.getSnapshotIndicator().getId());
-        toRet.setSnapshotEconomicId(originalSnapshot.getId());
-        toRet.setSurveyId(originalSnapshot.getSurveyDefinition().getId());
-        
-        //set family for information purpose
-        Long familyId = originalSnapshot.getFamily().getFamilyId();
-        toRet.setFamilyId(familyId);
-        toRet.setFamily(familyService.getFamilyById(familyId));
-        
         return toRet;
     }
 
@@ -270,16 +282,28 @@ public class SnapshotServiceImpl implements SnapshotService {
 
 	@Override
 	public void deleteSnapshotById(Long snapshotEconomicId) {
-		SnapshotEconomicEntity snapshotEconomicEntity = economicRepository.getOne(snapshotEconomicId);
-		SnapshotEconomicEntity snapshotEconomicEntityAux = snapshotEconomicEntity;
-		snapshotIndicatorPriorityRepository.delete(snapshotIndicatorPriorityRepository
-				.findBySnapshotIndicatorId(snapshotEconomicEntity.getSnapshotIndicator().getId()));
-		economicRepository.delete(snapshotEconomicEntity);
-		snapshotIndicatorRepository.delete(snapshotEconomicEntityAux.getSnapshotIndicator());
-		
-		if (economicRepository.findByFamilyFamilyId(snapshotEconomicEntityAux.getFamily().getFamilyId()).size() == 0) {
-			familyRepository.delete(snapshotEconomicEntityAux.getFamily().getFamilyId());
-		}
+    	SnapshotEconomicEntity snapshotEconomicEntity = economicRepository.
+    	        findOne(snapshotEconomicId);
+
+    	if (snapshotEconomicEntity!=null) {
+
+    	    Long familyId = snapshotEconomicEntity.getFamily().getFamilyId();
+
+            List<SnapshotIndicatorPriorityEntity> priorities =
+                    snapshotIndicatorPriorityRepository
+                    .findBySnapshotIndicatorId(snapshotEconomicEntity
+                            .getSnapshotIndicator().getId());
+
+        	if (priorities!=null && !priorities.isEmpty()) {
+        	    snapshotIndicatorPriorityRepository.delete(priorities);
+        	}
+
+        	economicRepository.delete(snapshotEconomicEntity);
+
+            if (economicRepository.findByFamilyFamilyId(familyId).size() == 0) {
+        	    familyRepository.delete(familyId);
+            }
+    	}
 	}
 
 }
